requirements.txt
alabaster==0.7.12
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-filter==2.2.0
django-modelsdoc==0.1.9
django-mysql==3.2.0
django-rest-swagger==2.2.0
django-sequences==2.4
django-templated-mail==1.1.1
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
docutils==0.15.2
drf-yasg==1.16.1
idna==2.8
imagesize==1.1.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
packaging==19.1
pydotplus==2.0.2
Pygments==2.4.2
pygraphviz==1.5
PyJWT==1.7.1
pykintone==0.3.10
pyparsing==2.4.2
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
